this half works. scrap.
pointers suck in c#. The Shader Parameter code works and writes the shader definitions. but the pointer idea needs the value to be fixed in memory to use it.

Maybe it would be a better idea to create an event for the value you want it to update to and subscribe to the event when it changes.

idk. im a bit upset this didnt work :(